+++ packages/spacecat-shared-data-access/src/models/site/config.js
@@ -60,7 +60,6 @@ export const Config = (data = {}) => {
   self.getSlackMentions = (type) => state?.handlers[type]?.mentions?.slack;
   self.getHandlerConfig = (type) => state?.handlers[type];
   self.getHandlers = () => state.handlers;
-  self.getHlxContentConfig = () => state.content;
   self.getImports = () => state.imports;
   self.getExcludedURLs = (type) => state?.handlers[type]?.excludedURLs;
   self.getManualOverwrites = (type) => state?.handlers[type]?.manualOverwrites;
@@ -72,13 +71,6 @@ export const Config = (data = {}) => {
       channel,
       workspace,
       invitedUserCount,
-    };
-  };
-
-  self.updateHlxContentConfig = (source, path) => {
-    state.content = {
-      source,
-      path,
     };
   };
 
@@ -118,7 +110,6 @@ Config.fromDynamoItem = (dynamoItem) => Config(dynamoItem);
 
 Config.toDynamoItem = (config) => ({
   slack: config.getSlackConfig(),
-  content: config.getHlxContentConfig(),
   handlers: config.getHandlers(),
   imports: config.getImports(),
 });